#183 move waypoints state into waypoints submodule
